+++ app/(route)/activity/edit/[id]/page.tsx
@@ -1,4 +1,3 @@
-import type { SubImage } from '@/_apis/activities/getActivity';
 import getActivity from '@/_apis/activities/getActivity';
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 
@@ -6,32 +5,6 @@ import CommonLayout from '@/_components/CommonLayout';
 import StickyLayout from '@/_components/SideStickyLayout';
 
 import EditLayout from './_components/EditLayout';
-import type { Schedule } from '../../register/page';
-
-export interface EditSchedule {
-  date: string;
-  endTime: string;
-  id?: number;
-  startTime: string;
-}
-
-export interface EditDetail {
-  schedules: EditSchedule[];
-  subImages: SubImage[];
-}
-
-export interface ActivityEdit {
-  address: string;
-  bannerImageUrl: string;
-  category: string;
-  description: string;
-  price: number | '';
-  scheduleIdsToRemove: number[];
-  schedulesToAdd: Schedule[];
-  subImageIdsToRemove: number[];
-  subImageUrlsToAdd: string[];
-  title: string;
-}
 
 /**
  * 체험 수정 페이지
