+++ app/(route)/activity/register/_components/IntroduceImage/index.tsx
@@ -3,13 +3,12 @@
 import type { Dispatch, SetStateAction } from 'react';
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
-import type { SubImage } from '@/_apis/activities/getActivity';
-import type { ActivityEdit, EditDetail } from '@/(route)/activity/edit/[id]/page';
+
+import type { Activity, ActivityEdit, EditDetail, SubImage } from '@/_types/activities/register';
 
 import { usePostImage } from '@/_hooks/use-post-image';
 import useModalState from '@/_hooks/useModalState';
 
-import type { Activity } from '../../page';
 import CommonModal from '../CommonModal';
 import FileInput from '../FileInput';
