+++ app/(route)/activity/register/page.tsx
@@ -4,24 +4,9 @@ import StickyLayout from '@/_components/SideStickyLayout';
 
 import RegisterLayout from './_components/RegisterLayout';
 
-export interface Schedule {
-  date: string;
-  endTime: string;
-  startTime: string;
-}
-
-export interface Activity {
-  address: string;
-  bannerImageUrl: string;
-  category: string;
-  description: string;
-  price: number | '';
-  schedules: Schedule[];
-  subImageUrls?: string[];
-  title: string;
-}
-
 function ActivityRegister() {
+  const isClient = typeof window !== 'undefined';
+
   return (
     <CommonLayout>
       <StickyLayout>
