+++ app/_apis/activities/getActivity.ts
@@ -1,25 +1,9 @@
 import type { AxiosResponse } from 'axios';
 import { AxiosError } from 'axios';
-import type { EditSchedule } from '@/(route)/activity/edit/[id]/page';
-import type { Activity } from '@/(route)/activity/register/page';
+
+import type { ActivityDetail } from '@/_types/activities/register';
 
 import axiosInstance from '@/_libs/axios';
-
-export interface SubImage {
-  id?: number;
-  imageUrl: string;
-}
-
-export interface ActivityDetail extends Activity {
-  createdAt: string;
-  id: number;
-  rating: number;
-  reviewCount: number;
-  schedules: EditSchedule[];
-  subImages: SubImage[];
-  updatedAt: string;
-  userId: number;
-}
 
 const getActivity = async (activityId: string) => {
   try {
